Fix auth closing activity
This solves a rare edge case when a user presses the home button instead
of authenticating. The app goes to background but the Activity is still
there because it was started as a new task. Now if user closes auth and
try opening app again, they'll open the old activity which destroys itself
as soon as it is opened. This commit makes sure that the activity dies if
the used stops the auth by going to homescreen. Ideally google should stop
BiometricPrompt as soon as activity stops.

¯\_(ツ)_/¯

Signed-off-by: Aditya Wasan <adityawasan55@gmail.com>